guardianGuide/backend/requirements.txt
fastapi==0.115.6
python-multipart==0.0.20
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic-settings==2.6.1
supabase==2.10.0
httpx==0.27.2
redis==5.2.1
langchain-google-genai==2.0.7
langchain-core==0.3.28
langgraph==0.2.59
duckduckgo-search==6.3.5
wikipedia==1.4.0
google-api-python-client==2.156.0
google-auth-oauthlib==1.2.1
twilio==9.3.7
resend==2.4.0
pytest==8.3.4
